requirements.txt
absl-py==0.9.0
alabaster==0.7.12
apispec==1.3.3
APScheduler==3.6.3
argh==0.26.2
asgiref==3.2.3
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
autovizwidget==0.13.1
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
blaze==0.11.3
bleach==3.1.0
bokeh==1.4.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.3.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
cytoolz==0.10.1
dask==2.10.1
datashape==0.5.4
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.10.0
dj-database-url==0.5.0
Django==3.0.3
django-filter==2.2.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
drfdocs==0.0.11
entrypoints==0.3
Fiona==1.8.4
flake8==3.7.9
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.0
Flask-APScheduler==1.11.0
Flask-Babel==0.12.2
Flask-Caching==1.7.1
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-JSON==0.3.4
Flask-JWT-Extended==3.20.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.2
fsspec==0.6.2
future==0.18.2
gast==0.2.2
GDAL==2.3.3
geopandas==0.6.1
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
hdijupyterutils==0.12.9
HeapDict==1.0.1
idna==2.8
imageio==2.8.0
imagesize==1.2.0
implicit==0.3.7
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
marshmallow==2.18.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.19.0
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
msgpack==0.6.1
multipledispatch==0.6.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
odo==0.5.1
olefile==0.46
openapi-codec==1.3.2
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.1
pandas-datareader==0.8.1
pandas-profiling==1.4.1
pandasql==0.7.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.0
pluggy==0.13.1
prison==0.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.6.7
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyproj==1.9.6
pyreadline==2.1
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==3.9.3
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
python-socketio==4.3.1
python3-openid==3.1.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
qgrid==1.1.1
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.16.0
rsa==4.0
Rtree==0.9.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
sqlparse==0.3.0
streamz==0.5.2
tables==3.6.1
tblib==1.6.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-addons==0.8.2
tensorflow-estimator==2.1.0
tensorflow-hub==0.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
typed-ast==1.4.1
typeguard==2.7.1
tzlocal==2.0.0
ujson==1.35
uritemplate==3.0.1
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
yapf==0.28.0
zict==1.0.0
zipp==2.2.0
